# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "APP SET V2",
        "description": "APP SET V2 Documentation",
        "contact": {
            "email": "alonso@setpay.mx"
        },
        "version": "2.0.0"
    },
    "paths": {
...
        "/api/cardCloud/card/{cardId}/sensitive": {
            "get": {
                "tags": [
                    "Card Cloud V2"
                ],
                "summary": "Obtener datos sensibles de la tarjeta",
                "description": "Obtener datos sensibles de la tarjeta",
                "operationId": "sensitive",
                "parameters": [
                    {
                        "name": "cardId",
                        "in": "path",
                        "description": "ID de la tarjeta",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Datos sensibles de la tarjeta",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "pan": {
                                            "type": "string",
                                            "example": "5161520880102992"
                                        },
                                        "expiration_date": {
                                            "type": "string",
                                            "example": "2029-08-01"
                                        },
                                        "pin": {
                                            "type": "string",
                                            "example": "1911"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "text/plain": {
                                "schema": {
                                    "type": "string",
                                    "example": "Unauthorized"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "text/plain": {
                                "schema": {
                                    "type": "string",
                                    "example": "Bad Request"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "content": {
                            "text/plain": {
                                "schema": {
                                    "type": "string",
                                    "example": "Not Found"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "content": {
                            "text/plain": {
                                "schema": {
                                    "type": "string",
                                    "example": "Internal Server Error"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/cardCloud/card/{cardId}/cvv": {
            "get": {
                "tags": [
                    "Card Cloud V2"
                ],
                "summary": "Obtener CVV dinámico de la tarjeta",
                "description": "Obtener CVV dinámico de la tarjeta",
                "operationId": "dynamicCvv",
                "parameters": [
                    {
                        "name": "cardId",
                        "in": "path",
                        "description": "ID de la tarjeta",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "CVV dinámico de la tarjeta",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "cvv": {
                                            "type": "string",
                                            "example": "275"
                                        },
                                        "expiration": {
                                            "type": "integer",
                                            "example": 1880236800
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "text/plain": {
                                "schema": {
                                    "type": "string",
                                    "example": "Unauthorized"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "text/plain": {
                                "schema": {
                                    "type": "string",
                                    "example": "Bad Request"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/cardCloud/movement/{uuid}": {
            "get": {
                "tags": [
                    "Card Cloud V2"
                ],
                "summary": "Show movement",
                "description": "Show movement",
                "operationId": "9f1385842d7cdeeb0cc98f2f7e12772d",
                "parameters": [
                    {
                        "name": "uuid",
                        "in": "path",
                        "description": "Movement UUID",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Movement",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "movement_id": {
                                            "type": "string",
                                            "example": "01935c81-abe0-7238-8635-46a8486259be"
                                        },
                                        "date": {
                                            "type": "integer",
                                            "example": 1732423822
                                        },
                                        "type": {
                                            "type": "string",
                                            "example": "PURCHASE"
                                        },
                                        "amount": {
                                            "type": "string",
                                            "example": "-429.00"
                                        },
                                        "balance": {
                                            "type": "string",
                                            "example": "325.73"
                                        },
                                        "authorization_code": {
                                            "type": "string",
                                            "example": "324411"
                                        },
                                        "description": {
                                            "type": "string",
                                            "example": "COMPRA EN TIENDA"
                                        },
                                        "status": {
                                            "type": "string",
                                            "example": "Approved"
                                        },
                                        "client_id": {
                                            "type": "string",
                                            "example": "SP0001275"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Error al obtener el movimiento",
                        "content": {
                            "text/plain": {
                                "schema": {
                                    "type": "string",
                                    "example": "Error al obtener el movimiento"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "text/plain": {
                                "schema": {
                                    "type": "string",
                                    "example": "Unauthorized"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
...
}